# docs/swagger.yaml
basePath: /
definitions:
  response.Response:
    properties:
      data:
        type: object
      message:
        type: string
      success:
        type: boolean
    type: object
  user.LoginParam:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.Register:
    properties:
      confirm_password:
        type: string
      email:
        example: google@gmail.com
        type: string
      password:
        type: string
      verify_code:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: 1coolluobo@gmail.com
    name: 1cool
    url: https://github.com/1cool/golang-web
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/1cool/golang-web
  title: github.com/1cool/golang-web Example API
  version: "1.0"
paths:
  /api/v1/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户注册
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 用户相关
  /api/v1/users/{id}:
    get:
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户详情
      tags:
      - 用户相关
  /api/v1/users/auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户登录
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.LoginParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 用户相关
swagger: "2.0"